move SystemMetrics enum
